refactor(file-exchange): move ArtifactStore into _token_store; _artifacts becomes shim
